artifacts/incidai/source/ai-models/requirements.txt
# Core Data Science
pandas==2.1.4
numpy==1.26.3
scipy==1.11.4

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0

# Traditional ML
scikit-learn==1.4.0
xgboost==2.0.3
lightgbm==4.2.0
imbalanced-learn==0.12.0

# Deep Learning
torch==2.1.2
transformers==4.36.2
sentence-transformers==2.3.1
datasets==2.16.1

# NLP & Embeddings
nltk==3.8.1
spacy==3.7.2
gensim==4.3.2

# Vector Search
faiss-cpu==1.7.4
chromadb==0.4.22

# LLM Integration
openai==1.7.2
langchain==0.1.1
langchain-community==0.0.13
tiktoken==0.5.2

# Utilities
joblib==1.3.2
tqdm==4.66.1
python-dotenv==1.0.0
pyyaml==6.0.1

# Model Interpretability
shap==0.44.0
lime==0.2.0.1

# Experiment Tracking
mlflow==2.9.2
wandb==0.16.2

# API & Deployment
fastapi==0.108.0
uvicorn==0.25.0
pydantic==2.5.3

# Testing
pytest==7.4.4
pytest-cov==4.1.0
